+++ shapemap/src/query_shape_map.rs
@@ -1,3 +1,5 @@
+use std::fmt::Display;
+
 use crate::{Association, NodeSelector, ShapeSelector};
 use prefixmap::PrefixMap;
 use serde_derive::Serialize;
@@ -51,5 +53,15 @@ impl QueryShapeMap {
         S: SRDF,
     {
         self.iter().flat_map(|assoc| assoc.iter_node_shape(rdf))
+    }
+}
+
+impl Display for QueryShapeMap {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            serde_json::to_string_pretty(&self).map_err(|_| std::fmt::Error)?
+        )
     }
 }